Add Onboarding

+++ app/controllers/concerns/authentication.rb
@@ -35,7 +35,11 @@ module Authentication
     end
 
     def after_authentication_url
-      session.delete(:return_to_after_authenticating) || root_url
+      if Current.user.needs_onboarding?
+        onboarding_path(:username)
+      else
+        session.delete(:return_to_after_authenticating) || root_url
+      end
     end
 
     def start_new_session_for(user)
